requirements.txt
# Core Dependencies
python-docx>=1.1.0
openpyxl>=3.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
PyPDF2>=3.0.0
PyYAML>=6.0.0

# Data & Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# API Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0

# Database
aiosqlite>=0.19.0

# Cache & Rate Limiting
redis>=5.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

# Development
pyinstaller>=5.0
jinja2>=3.1.0
